// server/apis/fm/clients/Post.ts
/**
* Generated by @proofgeist/fmdapi package
* 2022-07-12 03:17 pm -07:00
* https://github.com/proofgeist/fmdapi
* DO NOT EDIT THIS FILE DIRECTLY. Changes may be overritten
*/

import { z } from "zod";
import { DataApi } from "@proofgeist/fmdapi";
export const ZPost = z.object({
    "Id": z.string(),
    "CreationTimestamp": z.string(),
    "ModificationTimestamp": z.string(),
    "Slug": z.string(),
    "Type": z.string(),
    "Title": z.string(),
    "Body": z.string(),
    "PostId": z.string(),
    "Excerpt": z.string(),
    "AuthorId": z.string(),
    "FeatureImageUrl": z.string(),
    "AuthorName": z.string(),
    "AuthorAvatarUrl": z.string()
});
export type TPost = z.infer<typeof ZPost>;
if (!process.env.FM_DATABASE)
    throw new Error("Missing env var: FM_DATABASE");
if (!process.env.FM_SERVER)
    throw new Error("Missing env var: FM_SERVER");
if (!process.env.OTTO_API_KEY)
    throw new Error("Missing env var: OTTO_API_KEY");
export const client = DataApi<any, TPost>({
    auth: { apiKey: process.env.OTTO_API_KEY },
    db: process.env.FM_DATABASE,
    server: process.env.FM_SERVER,
    layout: "post_versions_web"
}, {
    fieldData: ZPost
});